+++ src/chatproviderservice/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [9.67.5](https://github.com/Quenty/NevermoreEngine/compare/@quenty/chatproviderservice@9.67.4...@quenty/chatproviderservice@9.67.5) (2026-07-30)
+
+**Note:** Version bump only for package @quenty/chatproviderservice
+
 ## [9.67.4](https://github.com/Quenty/NevermoreEngine/compare/@quenty/chatproviderservice@9.67.3...@quenty/chatproviderservice@9.67.4) (2026-07-29)
 
 ### Bug Fixes